requirements.txt
fastapi
uvicorn
streamlit
requests
python-dotenv
python-multipart    
groq